requirements.txt
torch>=2.0.0
optuna>=4.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
httpx>=0.24.0
openai>=1.0.0
feedparser>=6.0.0
python-dotenv>=1.0.0
vectorbt>=0.4.0
requests>=2.28.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
playwright>=1.50.0
scipy>=1.10.0
